version2/backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.1
networkx==3.2.1
pydantic==2.9.2
passlib[bcrypt]
python-jose
